requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
mysql-connector
Flask
dill

#-e .